Refactor parsing logic (#119)
Much of the parsing logic is procedural and lacks clear separation of concerns / abstraction layers making the code harder to read and harder to test.

These changes refactor the logic in the `parsers` directory and splits the logical concerns between an `html.py` file and a `parsers.py` file using an object-oriented approach.

The classes in `html.py` represent the HTML entities that are being parsed (like the `PlayByPlayPage` and `PlayByPlayTable`) where the classes have helper methods that abstract away the implementation details around how given pieces of data are scraped / acquired from a particular entity (i.e. the exact column index for a play's description).

The classes in `parsers.py` represent taking "raw" data from a given page / html entity and formatting the data contained within that entity.

In the future, the client itself will be exposed via a `class`, and thus, any relevant `parser` instances will be instantiated when the client is instantiated.

For now, for each API method, a new parser instance is instantiated. This slightly increases the overhead for calling these API methods, but the benefit of this additional readability outweighs the cost of this slight increase in logical overhead.