Regular Season Box Scores for a given player (#129)
Introduces an API for fetching regular season box scores for a given player.

The API takes a player identifier and season end year values and parses the associated `/gamelog` page (example: `https://www.basketball-reference.com/players/w/westbru01/gamelog/2019` for Russell Westbrook's 2018-19 season).

The returned values are the regular season box scores in the regular season box score table on the page that the player was active in (so the values do _not_ include games where the player was suspended or a `DNP`, etc).

Future work will add the ability to fetch playoff box scores as well as the ability to return all games, even where the player was inactive for some reason.

Includes the work in #114 by @ntsirakis.

Co-authored-by: jaebradley <jae.b.bradley@gmail.com>